+++ frontend/src/shared/lib/validators/isValidEmail.ts
@@ -1,0 +1,3 @@
+export function isValidEmail(value: string): boolean {
+  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value);
+}
